breaks GPU tests into two sets: math and feature
