+++ test_scripts/gpu/gpuPythonFeatureTests.py
@@ -1,0 +1,32 @@
+#   Copyright 2015,2016 Ufora Inc.
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+import unittest
+import pyfora
+import ufora.config.Setup as Setup
+import ufora.FORA.CUDA.GpuFeatureTestCases as GpuFeatureTestCases
+import ufora.FORA.CUDA.GpuTestUtil as GpuTestUtil
+import ufora.FORA.python.FORA as FORA
+
+class GpuPythonFeatureTests(
+        unittest.TestCase,
+        GpuTestUtil.GpuTestUtil,
+        GpuFeatureTestCases.GpuFeatureTestCases
+        ):
+    pass
+
+if __name__ == '__main__':
+    import ufora.config.Mainline as Mainline
+    Mainline.UnitTestMainline([FORA])
+
